pyproject.toml
[project]
name = "friday"
version = "0.1.0"
description = "Friday - your AI coding agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "rich>=14.1.0",
]

[project.scripts]
friday = "friday.main:main"

[tool.setuptools.packages.find]
where = ["src"]
